Streamlined Data Conversion implementation. Added conversions for Post and Resume data.

+++ PEngine.Core.Shared/Models/PostModel.cs
@@ -6,7 +6,7 @@ namespace PEngine.Core.Shared.Models
   public class PostModel : IGuidModel, ITimestampModel, IUniqueNameModel, ISubTitleModel
   {
     public Guid Guid { get; set; }
-    public Int32? LegacyID { get; set; }
+    public long? LegacyID { get; set; }
     public string Name { get; set; }
     public string Data { get; set; }
     public string IconFileName { get; set; }
